Adding Context path/rout in jboss config file (Spring Boot adaptation)

+++ src/main/java/spring/angular/demo/repository/RareRepository.java
@@ -1,15 +1,12 @@
 package spring.angular.demo.repository;
 
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
-import spring.angular.demo.entity.Movie;
 import spring.angular.demo.entity.MovieRate;
 
 @Repository
-public interface RareRepository extends CrudRepository<MovieRate, Long> {
+public interface RareRepository extends CrudRepository<MovieRate, Integer> {
 
 
 }
